+++ nodes/node3-driver-inputs/Controller/CHANGELOG.md
@@ -9,9 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 unreleased
 ## [1.0.3] - 2026-03-31
 - Updating file and function documentation in compliance with Software Standards and Guidelines V1.0.1 (SSG V1.0.1)
-- Updating variable names to be in compliance with SSG V1.0.1
+- Updating variable names to be in compliance with SSG V1.0.1 
 - Adding function prototypes and pin definitions to header file in compliance with SSG V1.0.1
-
+- Adding OperatingMode enum to header file, replacing 0x02,0x01,0x80 with MODE_STOPPED, MODE_OPERATIONAL,MODE_PREOP respectively
+- in function SendAllNMT(), NODE ID's have been updated to use variable names to aid readability
 
 unreleased
 ## [1.0.2] - 2026-03-24