+++ nodes/node3-driver-inputs/Controller/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+unreleased
+## [1.0.3] - 2026-04-10
+### Added 
+- Horn logic (sends sdo to horn on button 1 state change)
 
 unreleased
 ## [1.0.2] - 2026-03-24